Added filecaching, modified a parser
